65 add type hint to pyvoronoi

+++ pyvoronoi/voronoi.hpp
@@ -332,8 +332,8 @@ public:
 	std::vector<Point> GetPoints();
 	std::vector<Segment> GetSegments();
 
-    int CountPoints();
-    int CountSegments();
+    long long CountPoints();
+    long long CountSegments();
 	long long CountVertices();
 	long long CountEdges();
 	long long CountCells();
